+++ mobile/android/android-components/components/feature/session/src/main/java/mozilla/components/feature/session/SessionFeature.kt
@@ -15,7 +15,7 @@ import mozilla.components.support.base.feature.LifecycleAwareFeature
 class SessionFeature(
     private val sessionManager: SessionManager,
     private val goBackUseCase: SessionUseCases.GoBackUseCase,
-    engineView: EngineView,
+    private val engineView: EngineView,
     private val sessionId: String? = null
 ) : LifecycleAwareFeature, UserInteractionHandler {
     internal val presenter = EngineViewPresenter(sessionManager, engineView, sessionId)
@@ -47,7 +47,10 @@ class SessionFeature(
             sessionManager.findSessionById(it)
         } ?: sessionManager.selectedSession
 
-        if (session?.canGoBack == true) {
+        if (engineView.canClearSelection()) {
+            engineView.clearSelection()
+            return true
+        } else if (session?.canGoBack == true) {
             goBackUseCase(session)
             return true
         }